Remote logging added. All unwrap data is sent out the COM port on the radar node ("Scarecrow" app) and also sent across a radio link to a remote logging node ("Scarecrow_radio_monitor" app), every half second.

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Samraksh_eMote/SamrakshEmote_Samraksh_eMote_Algorithm__AcousticDetection.cpp
@@ -18,7 +18,7 @@
 using namespace Samraksh::eMote;
 static INT32 Pa = 1, Pm = 1;
 
-INT8 Algorithm_AcousticDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, INT32 param1, CLR_RT_TypedArray_double param2, INT8 param3, HRESULT &hr )
+UINT16 Algorithm_AcousticDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, INT32 param1, CLR_RT_TypedArray_double param2, INT8 param3, HRESULT &hr )
 {
     INT8 retVal = 0;
