+++ MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_eMote/Samraksh_eMote/NativeStubs/SamrakshEmote_Samraksh_eMote_Algorithm__RadarDetection.cpp
@@ -16,15 +16,15 @@
 
 using namespace Samraksh::eMote;
 
-INT8 Algorithm_RadarDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, CLR_RT_TypedArray_UINT16 param1, CLR_RT_TypedArray_UINT16 param2, INT32 param3, CLR_RT_TypedArray_INT16 param4, HRESULT &hr )
+UINT16 Algorithm_RadarDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, CLR_RT_TypedArray_UINT16 param1, CLR_RT_TypedArray_UINT16 param2, INT32 param3, CLR_RT_TypedArray_INT16 param4, HRESULT &hr )
 {
-    INT8 retVal = 0; 
+    UINT16 retVal = 0; 
     return retVal;
 }
 
-INT8 Algorithm_RadarDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, CLR_RT_TypedArray_UINT16 param1, INT32 param2, CLR_RT_TypedArray_INT16 param3, HRESULT &hr )
+UINT16 Algorithm_RadarDetection::DetectionCalculation( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT16 param0, CLR_RT_TypedArray_UINT16 param1, INT32 param2, CLR_RT_TypedArray_INT16 param3, HRESULT &hr )
 {
-    INT8 retVal = 0; 
+    UINT16 retVal = 0; 
     return retVal;
 }
